+++ src/main/java/duke/TaskManager.java
@@ -7,7 +7,8 @@ import duke.task.model.Deadline;
 import duke.task.model.Event;
 import duke.task.model.Task;
 import duke.task.model.ToDo;
-
+import duke.dukeUtility.parser.DateParser;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 
 import static duke.dukeUtility.parser.JsonTaskToObjectParser.*;
@@ -32,12 +33,12 @@ public class TaskManager {
         this._activeTasks.addTask(task);
         return task;
     }
-    public Event addNewEvent(String taskDescription, String from, String to) {
+    public Event addNewEvent(String taskDescription, LocalDateTime from, LocalDateTime to) {
         Event event = new Event(taskDescription, from, to, this.rollSerialNo(), false);
         this._activeTasks.addTask(event);
         return event;
     }
-    public Deadline addNewDeadline(String taskDescription, String deadlineString) {
+    public Deadline addNewDeadline(String taskDescription, LocalDateTime deadlineString) {
         Deadline deadLine = new Deadline(taskDescription, deadlineString, this.rollSerialNo(), false);
         this._activeTasks.addTask(deadLine);
         return deadLine;
@@ -105,14 +106,14 @@ public class TaskManager {
             Integer taskId = getJsonPropertyTaskId(jsonObj);
             String taskDescription = getJsonPropertyTaskDescription(jsonObj);
             Boolean done = getJsonPropertyDoneStatus(jsonObj);
-            String deadline = getJsonPropertyDeadline(jsonObj);
+            LocalDateTime deadline = getJsonPropertyDeadline(jsonObj);
             return new Deadline(taskDescription, deadline, taskId, done);
         } else if (isJsonTypeEvent(jsonObj)) {
             Integer taskId = getJsonPropertyTaskId(jsonObj);
             String taskDescription = getJsonPropertyTaskDescription(jsonObj);
             Boolean done = getJsonPropertyDoneStatus(jsonObj);
-            String from = getJsonPropertyFrom(jsonObj);
-            String to = getJsonPropertyTo(jsonObj);
+            LocalDateTime from = getJsonPropertyFrom(jsonObj);
+            LocalDateTime to = getJsonPropertyTo(jsonObj);
             return new Event(taskDescription, from, to, taskId, done);
         }
         throw new Exception("Json object not recognised as a task Object");