+++ src/main/java/duke/dukeUtility/parser/TaskToJsonParser.java
@@ -7,6 +7,7 @@ import duke.task.model.Event;
 import duke.task.model.ToDo;
 
 import static duke.dukeUtility.definition.TaskField.*;
+import static duke.dukeUtility.parser.DateParser.parseLocalDateTimeAsString;
 
 
 public class TaskToJsonParser extends Parser {
@@ -20,8 +21,8 @@ public class TaskToJsonParser extends Parser {
         obj.addProperty(TASK_FIELD_TASK_ID, event.getTaskId());
         obj.addProperty(TASK_FIELD_DESCRIPTION, event.getTaskDescription());
         obj.addProperty(TASK_FIELD_DONE_STATUS, event.isDone());
-        obj.addProperty(TASK_FIELD_FROM, event.getFrom());
-        obj.addProperty(TASK_FIELD_TO, event.getTo());
+        obj.addProperty(TASK_FIELD_FROM, parseLocalDateTimeAsString(event.getFrom()));
+        obj.addProperty(TASK_FIELD_TO, parseLocalDateTimeAsString(event.getTo()));
         return obj;
     }
 
@@ -31,7 +32,7 @@ public class TaskToJsonParser extends Parser {
         obj.addProperty(TASK_FIELD_TASK_ID, deadline.getTaskId());
         obj.addProperty(TASK_FIELD_DESCRIPTION, deadline.getTaskDescription());
         obj.addProperty(TASK_FIELD_DONE_STATUS, deadline.isDone());
-        obj.addProperty(TASK_FIELD_DEADLINE, deadline.getDeadline());
+        obj.addProperty(TASK_FIELD_DEADLINE, parseLocalDateTimeAsString(deadline.getDeadline()));
         return obj;
     }