+++ src/main/java/duke/task/model/Event.java
@@ -1,11 +1,14 @@
 package duke.task.model;
 
+import java.time.LocalDateTime;
+
+import static duke.dukeUtility.parser.DateParser.prettifyLocalDateTime;
+
 public final class Event extends Task {
-    private String _from;
-    private String _to;
+    private LocalDateTime _from;
+    private LocalDateTime _to;
 
-
-    public Event(String taskDescription, String from, String to, Integer taskId, Boolean done) {
+    public Event(String taskDescription, LocalDateTime from, LocalDateTime to, Integer taskId, Boolean done) {
         super(taskDescription, taskId, done);
         this.setFrom(from);
         this.setTo(to);
@@ -14,24 +17,26 @@ public final class Event extends Task {
     private Event() {
     }
 
-    public String getTo() {
+    public LocalDateTime getTo() {
         return this._to;
     }
 
-    public String getFrom() {
+    public LocalDateTime getFrom() {
         return this._from;
     }
 
-    private void setFrom(String from) {
+    private void setFrom(LocalDateTime from) {
         this._from = from;
     }
 
-    private void setTo(String to) {
+    private void setTo(LocalDateTime to) {
         this._to = to;
     }
 
     public String getChronologyString() {
-        return String.format("From: %s, To: %s", this.getFrom(), this.getTo());
+
+        return String.format("From: %s, To: %s", prettifyLocalDateTime(this.getFrom()), prettifyLocalDateTime(this.getTo()));
+
     }
 
 }