+++ src/test/java/duke/TaskManager_Test.java
@@ -2,7 +2,10 @@ package duke;
 
 import duke.task.model.Event;
 import duke.testHelper.TestStream;
+import duke.testHelper.help.ParserUnderTest;
 import org.junit.jupiter.api.Test;
+
+import java.time.LocalDateTime;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -23,16 +26,17 @@ public class TaskManager_Test extends TestStream {
 
 
     @Test
-    public void Test_AddEvents(){
+    public void Test_AddEvents() throws Exception{
         TaskManager tm = new TaskManager();
 
         String desc0 = "desc0";
         String desc1 = "desc1";
-        tm.addNewEvent(desc0,"from","to");
+        tm.addNewEvent(desc0,ParserUnderTest.parseStringAsLocalDateTime("20210101"),ParserUnderTest.parseStringAsLocalDateTime("20210101"));
 
-        String from1 = "from1";
-        String to1 = "to1";
+        LocalDateTime from1 = ParserUnderTest.parseStringAsLocalDateTime("20210101");
+        LocalDateTime to1 = ParserUnderTest.parseStringAsLocalDateTime("20210101");
         tm.addNewEvent(desc1,from1,to1);
+
         assertEquals(desc0, tm.getTaskById(0).getTaskDescription());
         assertEquals(desc1, tm.getTaskById(1).getTaskDescription());