+++ src/test/java/duke/level/Increment_04_Test.java
@@ -8,6 +8,7 @@ import duke.mock.mockTask.MockEvent;
 import duke.mock.mockTask.MockTask;
 import duke.mock.mockTask.MockToDo;
 import duke.testHelper.TestStream;
+import duke.testHelper.help.ParserUnderTest;
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
@@ -107,8 +108,8 @@ public class Increment_04_Test extends TestStream {
         String taskDesc0 = "event_desc_abc asfasfasf";
 
 
-        String fromDateString = "fromday";
-        String toDateString = "today";
+        String fromDateString = "19990101";
+        String toDateString = "19990202";
         String storeEventCommand0 = generateTextCommandLineAddEvent(PROMPT_UNDER_TEST_ADD_EVENT, taskDesc0,DELIMITER_EVENT_FROM,fromDateString,DELIMITER_EVENT_TO,toDateString);
         String listCommand = generateTextCommandList(PROMPT_UNDER_TEST_LIST);
         String exitCommand = generateTextCommandExit(PROMPT_UNDER_TEST_EXIT_LOOP);
@@ -130,13 +131,13 @@ public class Increment_04_Test extends TestStream {
         StringBuilder expectedResponseBuilder = new StringBuilder();
         TaskManager tm = new TaskManager();
         FileResourceManager frm = new FileResourceManager(getDefaultTasksTestExportPathString(),getDefaultTasksImportTestPathString());
-        MockEvent expectedEvent = new MockEvent(taskDesc0, 0, false, fromDateString,toDateString);
+        MockEvent expectedEvent = new MockEvent(taskDesc0, 0, false, ParserUnderTest.parseStringAsLocalDateTime(fromDateString),ParserUnderTest.parseStringAsLocalDateTime(toDateString));
 
         expectedResponseBuilder.append(getMsgUnderTestEntry());
         expectedResponseBuilder.append(getMsgUnderTestAttemptImport(frm.getImportPath()));
         expectedResponseBuilder.append(getMsgUnderTestReadPathNotFound());
         expectedResponseBuilder.append(getMsgUnderTestBeginInputLoop());
-        expectedResponseBuilder.append(getMsgUnderTestResponseEventAdded(taskDesc0));
+        expectedResponseBuilder.append(getMsgUnderTestResponseEventAdded(expectedEvent.getDesc()));
 
         MockTask[] MockEvents = {expectedEvent};
 
@@ -165,9 +166,8 @@ public class Increment_04_Test extends TestStream {
         StringBuilder commandBuilder = new StringBuilder();
 
         String taskDesc0 = "event_desc_abc asfasfasf";
+        String byDateString = "20200101";
 
-
-        String byDateString = "midnight";
         String storeDeadlineCommand0 = generateTextCommandLineAddDeadline(PROMPT_UNDER_TEST_ADD_DEADLINE, taskDesc0,DELIMITER_DEADLINE_DEADLINE,byDateString);
         String listCommand = generateTextCommandList(PROMPT_UNDER_TEST_LIST);
         String exitCommand = generateTextCommandExit(PROMPT_UNDER_TEST_EXIT_LOOP);
@@ -192,7 +192,7 @@ public class Increment_04_Test extends TestStream {
         FileResourceManager frm = new FileResourceManager(getDefaultTasksTestExportPathString(),getDefaultTasksImportTestPathString());
 
 
-        MockDeadline expectedEvent = new MockDeadline(taskDesc0, 0, false, byDateString);
+        MockDeadline expectedEvent = new MockDeadline(taskDesc0, 0, false, ParserUnderTest.parseStringAsLocalDateTime(byDateString));
 
         expectedResponseBuilder.append(getMsgUnderTestEntry());
         expectedResponseBuilder.append(getMsgUnderTestAttemptImport(frm.getImportPath()));
@@ -249,10 +249,10 @@ public class Increment_04_Test extends TestStream {
     public void TestLevel4_Greet_AddEachEvent_List_Exit() throws Exception{
         String task0ToDoDescription = "todo_desc asfasfasf";
         String task1DeadlineDescription = "deadline_desc ndfrgndfndfn";
-        String task1DeadlineByString = "someDeadline";
+        String task1DeadlineByString = "20010101";
         String task2EventDescription = "event_desc_abc 213t12 3b52";
-        String task2EventFromDateString = "fromday";
-        String task2EventToDateString = "today";
+        String task2EventFromDateString = "20200102";
+        String task2EventToDateString = "20200102";
 
         String storeToDoCommand0 = generateTextCommandLineAddToDo(PROMPT_UNDER_TEST_ADD_TO_DO, task0ToDoDescription);
         String storeDeadlineCommand1 = generateTextCommandLineAddDeadline(PROMPT_UNDER_TEST_ADD_DEADLINE, task1DeadlineDescription,DELIMITER_DEADLINE_DEADLINE,task1DeadlineByString);
@@ -288,8 +288,9 @@ public class Increment_04_Test extends TestStream {
         StringBuilder expectedResponseBuilder = new StringBuilder();
 
         MockToDo expectedToDo = new MockToDo(task0ToDoDescription, 0,false);
-        MockDeadline expectedDeadline = new MockDeadline(task1DeadlineDescription, 1, false, task1DeadlineByString);
-        MockEvent expectedEvent = new MockEvent(task2EventDescription, 2, false, task2EventFromDateString,task2EventToDateString);
+
+        MockDeadline expectedDeadline = new MockDeadline(task1DeadlineDescription, 1, false, ParserUnderTest.parseStringAsLocalDateTime(task1DeadlineByString));
+        MockEvent expectedEvent = new MockEvent(task2EventDescription, 2, false, ParserUnderTest.parseStringAsLocalDateTime(task2EventFromDateString),ParserUnderTest.parseStringAsLocalDateTime(task2EventToDateString));
 
         TaskManager tm = new TaskManager();
         FileResourceManager frm = new FileResourceManager(getDefaultTasksTestExportPathString(),getDefaultTasksImportTestPathString());